Automatically drag all active (selected) files
By replacing the button widget with a toggle button, the files can be
flexibly dragged based on the selected status. It offers a middle ground
between the all or one options. Secondly, the receiver may limit the
amount of files, thus negating the need for closing and re-opening the
application with new files.