Add shortcut keys and begin work on a command bar
Some receivers have limits on how many items may be dropped in a single
operation. Introduce `selectable_next` which will select the next X
files in the group with 'X' being the current number of selected items.

For now hardcode the 'j' key to `selectable_next` to emulate vi's motion
key. Map the 'a' and 'd' keys to respectively select and deselect all
items.

Reparent the scroll panel to a grid and share it with an action bar
widget. The goal is to have buttons represent the previously mentioned
actions. The action bar should enabled state should be controlled a CLI
switch corresponding to the `command_bar_enabled` boolean.